bugfix: the generated code was not kept when the shell option keepdebug was supplied

+++ engine-net-2/src/lgspBackend/lgspGrGen.cs
@@ -2132,7 +2132,8 @@ namespace de.unika.ipd.grGen.lgsp
             }
             finally
             {
-                Directory.Delete(dirname, true);
+                if((flags & ProcessSpecFlags.KeepGeneratedFiles) == 0 && (flags & ProcessSpecFlags.UseExistingMask) == ProcessSpecFlags.UseNoExistingFiles)
+                    Directory.Delete(dirname, true);
             }
         }
     }
